requirements.txt
pyyaml>=6.0
jsonschema>=4.20
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
gitpython>=3.1
jinja2>=3.1
